Updated folder layout so its more modular, also update README.md correctly
